etl/requirements.txt
requests
pandas
sqlalchemy
pymysql
python-dotenv
cryptography
pytest
pytest-cov
